+++ pdf/lib/src/pdf/exif.dart
@@ -17,7 +17,7 @@
 import 'dart:convert';
 import 'dart:typed_data';
 
-import 'image.dart';
+import 'obj/image.dart';
 
 /// Jpeg metadata extraction
 class PdfJpegInfo {
