+++ pdf/lib/src/pdf/font/font_metrics.dart
@@ -18,7 +18,7 @@ import 'dart:math' as math;
 
 import 'package:meta/meta.dart';
 
-import 'rect.dart';
+import '../rect.dart';
 
 /// Describe dimensions for glyphs in a font
 @immutable
