+++ pdf/lib/src/pdf/graphic_state.dart
@@ -20,9 +20,9 @@ import 'package:meta/meta.dart';
 
 import 'data_types.dart';
 import 'document.dart';
-import 'function.dart';
-import 'object_dict.dart';
-import 'smask.dart';
+import 'obj/function.dart';
+import 'obj/object_dict.dart';
+import 'obj/smask.dart';
 
 enum PdfBlendMode {
   /// Selects the source colour, ignoring the backdrop
