+++ pdf/pubspec.yaml
@@ -3,7 +3,7 @@ description: A pdf producer for Dart. It can create pdf files for both web or fl
 homepage: https://github.com/DavBfr/dart_pdf/tree/master/pdf
 repository: https://github.com/DavBfr/dart_pdf
 issue_tracker: https://github.com/DavBfr/dart_pdf/issues
-version: 3.6.6
+version: 3.7.0
 
 environment:
   sdk: ">=2.12.0 <3.0.0"
