+++ pdf/test/arabic_test.dart
@@ -16,7 +16,7 @@
 
 import 'dart:io';
 
-import 'package:pdf/src/pdf/arabic.dart' as arabic;
+import 'package:pdf/src/pdf/font/arabic.dart' as arabic;
 import 'package:pdf/widgets.dart';
 import 'package:test/test.dart';
